requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
anyio==4.9.0
    # via httpx
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via python-telegram-bot
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
python-telegram-bot==22.0
    # via test-render (pyproject.toml)
requests==2.32.3
    # via test-render (pyproject.toml)
sniffio==1.3.1
    # via anyio
urllib3==2.3.0
    # via requests
